@@ -12,7 +12,7 @@ from .response import BaseHTTPResponse
 __all__ = ["RequestMethods"]
 
 _TYPE_ENCODE_URL_FIELDS = typing.Union[
-    typing.Sequence[typing.Tuple[str, typing.Union[str, bytes]]],
+    typing.Sequence[tuple[str, typing.Union[str, bytes]]],
     typing.Mapping[str, typing.Union[str, bytes]],
 ]
 
