@@ -7,9 +7,9 @@ import email.utils
 import gzip
 import mimetypes
 import zlib
+from collections.abc import Iterator
 from io import BytesIO
 from pathlib import Path
-from typing import Iterator
 
 import trio
 from quart import Response, make_response, request
